requirements.txt
Flask==3.1.1
google-generativeai==0.8.5
elevenlabs==2.24.0
numpy==1.26.4
python-dotenv==1.0.0
